+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added `\Wizaplace\SDK\Order\DeclinationOption`
 - Added `\Wizaplace\SDK\Order\OrderItem::getDeclinationOptions`
 - `\Wizaplace\SDK\User\UpdateUserCommand::setBirthday` now takes a `\DateTimeInterface` instead of a `\DateTime`
+- Added `\Wizaplace\SDK\Catalog\ProductAttribute::getId`
 
 ### Bugfixes
 </details>